requirements.txt
fastapi==0.110.0
uvicorn==0.27.1

requests==2.31.0

python-dotenv==1.0.1

tenacity==8.2.3
google-generativeai>=0.4.1
